// mobile/src/types/glossy.ts
export type TradeCategory =
  | 'painting-decorating'
  | 'plastering'
  | 'flooring'
  | 'plumbing'
  | 'electrical'
  | 'tiling'
  | 'carpentry'
  | 'kitchen-fitting'
  | 'bathroom-fitting'
  | 'handyman';